IMDBSearch-main/requirements.txt
django~=4.1.4
pathlib~=1.0.1
pika~=1.3.1
djangorestframework~=3.14.0
pandas~=1.4.4
mysqlclient~=2.0.3
redis==3.5.3
django_redis~=5.0.0
django-cluster-redis==1.0.5
redis-py-cluster==2.1.3